+++ presentation/src/main/java/com/lighthouse/presentation/binding/TextView.kt
@@ -1,4 +1,4 @@
-package com.lighthouse.presentation.bindingadapter
+package com.lighthouse.presentation.binding
 
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
